refactor: :recycle: separate image functions into utils module
